+++ Runtime/Scripts/FakeSchema/Root.cs
@@ -1,0 +1,11 @@
+﻿namespace GLTFast.FakeSchema {
+    
+    [System.Serializable]
+    public class Root
+    {
+        /// <summary>
+        /// An array of materials. A material defines the appearance of a primitive.
+        /// </summary>
+        public Material[] materials;
+    }
+}