+++ Runtime/Scripts/FakeSchema/RootChild.cs
@@ -1,0 +1,8 @@
+﻿
+namespace GLTFast.FakeSchema {
+
+    [System.Serializable]
+    public class RootChild {
+        public string name;        
+    }
+}