+++ Runtime/Scripts/Schema/Material.cs
@@ -15,6 +15,9 @@ namespace GLTFast.Schema {
             BLEND
         }
 
+        public MaterialExtension extensions;
+
+
         /// <summary>
         /// A set of parameter values that are used to define the metallic-roughness
         /// material model from Physically-Based Rendering (PBR) methodology.