+++ Runtime/Scripts/Schema/MaterialExtension.cs
@@ -1,0 +1,8 @@
+namespace GLTFast.Schema {
+
+    [System.Serializable]
+    public class MaterialExtension {
+        public PbrSpecularGlossiness KHR_materials_pbrSpecularGlossiness;
+        public MaterialUnlit KHR_materials_unlit;
+    }
+}