+++ Runtime/Scripts/Schema/MaterialUnlit.cs
@@ -1,0 +1,5 @@
+namespace GLTFast.Schema {
+    [System.Serializable]
+    public class MaterialUnlit {
+    }
+}